backend/backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
cloudinary==1.44.0
Django==5.2.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
idna==3.10
pillow==11.2.1
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
razorpay==1.4.2
reportlab==4.4.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
